+++ plume/src/test/kotlin/io/github/plume/oss/domain/DeltaGraphTest.kt
@@ -20,6 +20,7 @@ import org.junit.jupiter.api.Test
 import overflowdb.Config
 import overflowdb.Graph
 
+@Suppress("DEPRECATION")
 class DeltaGraphTest {
 
     companion object {